Tombstones are now created during populate phase.
Previously only new data was inserted which was inconsistent with the --deletes flag.
tlp-stress now honors the --deletes flag during populate operations.

Additionally, the RandomPartitionAccess workload has a new parameter, delete, which
behaves the same way as the select parameter but allows the user to delete
an entire partition at a time vs having the same behavior apply to both SELECT
and DELETE statements.