Renamed TableScan to RangeScan and added option to use token splits.
This workload can be used to perform either full table or range scans,
making it a good option to test virtual tables or to test the performance
of range scans against an existing table.

It doesn't include it's own table, you must pass the table name
as a workload parameter.